KAFKA-14966; [3/N] Extract reusable functionality for validating positions
