+++ clients/src/main/java/org/apache/kafka/clients/consumer/internals/OffsetFetcherUtils.java
@@ -18,6 +18,8 @@ package org.apache.kafka.clients.consumer.internals;
 
 import org.apache.kafka.clients.ApiVersions;
 import org.apache.kafka.clients.NodeApiVersions;
+import org.apache.kafka.clients.consumer.LogTruncationException;
+import org.apache.kafka.clients.consumer.OffsetAndMetadata;
 import org.apache.kafka.clients.consumer.OffsetAndTimestamp;
 import org.apache.kafka.clients.consumer.OffsetResetStrategy;
 import org.apache.kafka.common.IsolationLevel;
@@ -37,9 +39,11 @@ import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.common.utils.Time;
 import org.slf4j.Logger;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
@@ -313,6 +317,58 @@ class OffsetFetcherUtils {
         if (!(error instanceof RetriableException) && !cachedListOffsetsException.compareAndSet(null,
                 error))
             log.error("Discarding error in ListOffsetResponse because another error is pending", error);
+    }
+
+
+    void onSuccessfulRequestForValidatingPositions(
+            final Map<TopicPartition, SubscriptionState.FetchPosition> fetchPositions,
+            final OffsetsForLeaderEpochUtils.OffsetForEpochResult offsetsResult) {
+        List<SubscriptionState.LogTruncation> truncations = new ArrayList<>();
+        if (!offsetsResult.partitionsToRetry().isEmpty()) {
+            subscriptionState.setNextAllowedRetry(offsetsResult.partitionsToRetry(),
+                    time.milliseconds() + retryBackoffMs);
+            metadata.requestUpdate();
+        }
+
+        // For each OffsetsForLeader response, check if the end-offset is lower than our current offset
+        // for the partition. If so, it means we have experienced log truncation and need to reposition
+        // that partition's offset.
+        // In addition, check whether the returned offset and epoch are valid. If not, then we should reset
+        // its offset if reset policy is configured, or throw out of range exception.
+        offsetsResult.endOffsets().forEach((topicPartition, respEndOffset) -> {
+            SubscriptionState.FetchPosition requestPosition = fetchPositions.get(topicPartition);
+            Optional<SubscriptionState.LogTruncation> truncationOpt =
+                    subscriptionState.maybeCompleteValidation(topicPartition, requestPosition,
+                            respEndOffset);
+            truncationOpt.ifPresent(truncations::add);
+        });
+
+        if (!truncations.isEmpty()) {
+            maybeSetOffsetForLeaderException(buildLogTruncationException(truncations));
+        }
+    }
+
+    void onFailedRequestForValidatingPositions(final Map<TopicPartition,
+            SubscriptionState.FetchPosition> fetchPositions,
+                                               final RuntimeException error) {
+        subscriptionState.requestFailed(fetchPositions.keySet(), time.milliseconds() + retryBackoffMs);
+        metadata.requestUpdate();
+
+        if (!(error instanceof RetriableException)) {
+            maybeSetOffsetForLeaderException(error);
+        }
+    }
+
+    private LogTruncationException buildLogTruncationException(List<SubscriptionState.LogTruncation> truncations) {
+        Map<TopicPartition, OffsetAndMetadata> divergentOffsets = new HashMap<>();
+        Map<TopicPartition, Long> truncatedFetchOffsets = new HashMap<>();
+        for (SubscriptionState.LogTruncation truncation : truncations) {
+            truncation.divergentOffsetOpt.ifPresent(divergentOffset ->
+                    divergentOffsets.put(truncation.topicPartition, divergentOffset));
+            truncatedFetchOffsets.put(truncation.topicPartition, truncation.fetchPosition.offset);
+        }
+        return new LogTruncationException("Detected truncated partitions: " + truncations,
+                truncatedFetchOffsets, divergentOffsets);
     }
 
     // Visible for testing
